fix(BJS2-40530): jacoco classDirectories
added classDirectories to jacocoTestCoverageVerification in build.gradle.kts
