+++ src/test/java/school/faang/user_service/service/filters/UserEmailFilterTest.java
@@ -1,23 +1,23 @@
-package school.faang.user_service.service.filteres;
+package school.faang.user_service.service.filters;
 
 import org.junit.jupiter.api.Test;
 import school.faang.user_service.dto.UserFilterDto;
 import school.faang.user_service.entity.User;
-import school.faang.user_service.filters.UserNameFilter;
+import school.faang.user_service.filters.UserEmailFilter;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-class UserNameFilterTest {
-    final UserNameFilter filter = new UserNameFilter();
+class UserEmailFilterTest {
+    final UserEmailFilter filter = new UserEmailFilter();
 
     @Test
     void testMatches() {
         User user = new User();
-        user.setUsername("Vlad");
+        user.setEmail("test@mail");
 
         UserFilterDto filterDto = new UserFilterDto();
-        filterDto.setNamePattern("Vl(.*)");
+        filterDto.setEmailPattern("(.*)@mail");
 
         assertTrue(filter.apply(user, filterDto));
     }
@@ -25,10 +25,10 @@ class UserNameFilterTest {
     @Test
     void testNotMatches() {
         User user = new User();
-        user.setUsername("Ivan");
+        user.setEmail("test@gmail");
 
         UserFilterDto filterDto = new UserFilterDto();
-        filterDto.setNamePattern("Vl(.*)");
+        filterDto.setEmailPattern("(.*)@mail");
 
         assertFalse(filter.apply(user, filterDto));
     }
