+++ src/test/java/school/faang/user_service/service/filters/event/EventUserIdFilterTest.java
@@ -1,8 +1,9 @@
-package school.faang.user_service.service.filteres.event;
+package school.faang.user_service.service.filters.event;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import school.faang.user_service.dto.event.EventFilterDto;
 import school.faang.user_service.entity.User;
@@ -16,7 +17,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @ExtendWith(MockitoExtension.class)
 public class EventUserIdFilterTest {
-    private final EventUserIdFilter eventUserIdFilter = new EventUserIdFilter();
+    @Spy
+    private EventUserIdFilter eventUserIdFilter;
     private Stream<Event> events;
     private EventFilterDto filterDto;
